Massive Restructuring Of Dataset Files
If I'm going to work on a project, the folders have to make sense. They just, have to. Some of them did, some of them did not. Assets is a better file structure than not, because you know what type of file should be in any subsequent folder making up the difference. As opposed to a bunch of redundant categories housing various files not dedicated to being what they are, under one unified purpose.